+++ src/black_primer/lib.py
@@ -120,7 +120,7 @@ async def black_run(
     """Run Black and record failures"""
     cmd = [str(which(BLACK_BINARY))]
     if "cli_arguments" in project_config and project_config["cli_arguments"]:
-        cmd.extend(*project_config["cli_arguments"])
+        cmd.extend(project_config["cli_arguments"])
     cmd.append("--check")
     if no_diff:
         cmd.append(".")