+++ CHANGES.md
@@ -1,14 +1,16 @@
 ## Change Log
 
-### Unreleased
+### 21.5b0
 
 #### _Black_
 
 - Set `--pyi` mode if `--stdin-filename` ends in `.pyi` (#2169)
-- Add `--no-diff` to black-primer to suppress formatting changes (#2187)
-
 - Stop detecting target version as Python 3.9+ with pre-PEP-614 decorators that are
   being called but with no arguments (#2182)
+
+#### _Black-Primer_
+
+- Add `--no-diff` to black-primer to suppress formatting changes (#2187)
 
 ### 21.4b2
 
@@ -19,6 +21,9 @@
 - Clarify
   [circumstances](https://github.com/psf/black/blob/master/docs/the_black_code_style.md#pragmatism)
   in which _Black_ may change the AST (#2159)
+
+- Allow `.gitignore` rules to be overridden by specifying `exclude` in `pyproject.toml`
+  or on the command line. (#2170)
 
 #### _Packaging_